Move the generic nodejs Process::Status implementation to corelib
Also remove requiring "native" which is not used here.
